things kinda work now

+++ sims/AstraSim/launch_gcp.py
@@ -315,15 +315,16 @@ def run_task(task):
     else:
         raise NotImplementedError
 
-    # os.system(cmd)
-
-    process = subprocess.Popen(cmd, shell=True)
-    timeout_sec = float(task["timeout"])
-    try:
-        process.wait(timeout=timeout_sec)
-    except subprocess.TimeoutExpired:
-        print(f"Timeout expired after {timeout_sec} seconds. Terminating the process...")
-        process.terminate()
+    if algo == "bo":
+        process = subprocess.Popen(cmd, shell=True)
+        timeout_sec = float(task["timeout"])
+        try:
+            process.wait(timeout=timeout_sec)
+        except subprocess.TimeoutExpired:
+            print(f"Timeout expired after {timeout_sec} seconds. Terminating the process...")
+            process.terminate()
+    else:
+        os.system(cmd)
 
 
 def main(_):
